// src/api/generated/models/additionalCourseResource.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * brew
 * OpenAPI spec version: 0.0.1
 */
import type { AdditionalCourseResourceCertificate } from './additionalCourseResourceCertificate';
import type { AdditionalCourseResourceCreatedAt } from './additionalCourseResourceCreatedAt';
import type { AdditionalCourseResourceUpdatedAt } from './additionalCourseResourceUpdatedAt';

export interface AdditionalCourseResource {
  id: string;
  name: string;
  provider: string;
  dev_profile_id: string;
  verified: boolean;
  certificate: AdditionalCourseResourceCertificate;
  created_at: AdditionalCourseResourceCreatedAt;
  updated_at: AdditionalCourseResourceUpdatedAt;
}